pciutils: fix scan-build warnings
The scan-build utility found some code that it didn't like, and this
patch fixes the small issues found.

The first ones in lib/access.c are mostly because scan-build doesn't
like the __attribute((noreturn)) on a function that gets inlined
in a function that has return statements.  This is undefined
behavior from the C-specification as I understand it.

The fix itself of initializing stack variables to zero adds very little
run-time code, but makes sure the heap isn't inputting odd values into
the function, and makes the static analysis pass.

The second fix is quite a bit more involved in understanding the steps
that lead to the code being plausibly wrong, but at the core of it,
scan-build can't quite understand the complexities of the library calls
to string libraries and that there was an implicit (not visible in code)
dependency on the values of 'value' and the values of the 'n' variable.

Simply adding a check on the 'value' variable to make sure it's non-null
when looping over the characters in the string solves the implicit
dependency and doesn't change the code flow.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>